requirements.txt
requests
openai
langchain==0.0.142
python-dotenv
